fix wiki property searching/updates
This change attempts to clarify a lot of the code while updating some
of the logic around how wiki properties are chosen and updated. A few
changes:
- numbers are written in ‘raw’ format without thousands separator
- references are added to the edits for county population
- merging some of the logic between states and counties, instead of
commenting out the state version

Many of these changes are bundled together because they were dependent
on changing the dict structure used for searching for wiki values to
update (‘total_pop - 1’ —> ‘pop_with_ref’)

Note: While this cleans up some of the logic about which combination of
wiki properties are associated with each other, we’re not adding new
keys yet when a necessary one is missing.