be explicit about how wiki templates are updated
This largely changes the logic for updating properties within the Wiki
templates. The problem is that each template expects values to be
formatted in a particular way, and each template behaves differently
depending on which properties are present. For example, ‘Infobox
settlement’ supports displaying both the decennial population as well
as an estimate. However, ‘Infobox U.S. County’ uses the same property
for both, but only displays the estimate if the decennial year isn’t
provided in a separate property. This approach also consolidates the
template logic instead of spreading its assumptions throughout the
wikibot code.

For logging, we’ll display a full diff of all text that will be
changed, instead of logging each property individually. This helps to
identify how whitespace is managed, as well as how the properties are
integrated into their surrounding context.

The template manipulation is also updated to use mwparserfromhell,
instead of raw string manipulation.